# api_spec.yaml
openapi: 3.0.0
info:
  title: DART Analytics API
  description: |
    Professional REST API for generating MSD Control Charts with statistical analysis.
    
    ## Features
    - Multi-format data upload (CSV, Excel)
    - Advanced statistical analysis
    - PowerPoint export
    - Real-time chart generation
    
    ## Authentication
    Currently, no authentication is required (suitable for internal use).
    For production deployment, consider adding API keys or OAuth2.
  version: 2.0.0
  contact:
    name: DART Analytics Team
    email: support@dartanalytics.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:5000
    description: Local development server
  - url: http://your-production-url.com
    description: Production server

tags:
  - name: File Operations
    description: Upload and manage data files
  - name: Chart Generation
    description: Generate control charts with analysis
  - name: Export
    description: Export reports to various formats
  - name: System
    description: System health and monitoring

paths:
  /api/upload:
    post:
      tags:
        - File Operations
      summary: Upload data file
      description: |
        Upload a CSV or Excel file for analysis. The system automatically:
        - Detects file encoding
        - Identifies column types (numeric, date, categorical)
        - Validates data quality
        - Creates a session for multi-chart analysis
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV or Excel file (.csv, .xlsx, .xls)
              required:
                - file
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "File uploaded successfully"
                  session_id:
                    type: string
                    example: "1234567890123"
                  columns:
                    type: object
                    properties:
                      numeric:
                        type: array
                        items:
                          type: string
                      date_like:
                        type: array
                        items:
                          type: string
                      categorical:
                        type: array
                        items:
                          type: string
                  data_info:
                    type: object
                    properties:
                      rows:
                        type: integer
                      columns:
                        type: integer
                      missing_values:
                        type: integer
        '400':
          description: Bad request (invalid file or format)
        '500':
          description: Server error

  /api/generate_chart:
    post:
      tags:
        - Chart Generation
      summary: Generate MSD control chart
      description: |
        Generate a Moving Standard Deviation (MSD) control chart with statistical analysis.
        
        ## Parameters
        - **date_col**: Time series column(s) - can be single or multiple columns
        - **value_col**: Numeric column to analyze
        - **group_col**: Optional grouping column for separate analysis
        - **rolling_window**: Window size for moving statistics (default: 10)
        - **std_multiplier**: Standard deviation multiplier for control limits (default: 3)
        - **resample**: Time aggregation (D=daily, W=weekly, M=monthly, Q=quarterly, Y=yearly)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  example: "1234567890123"
                date_col:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
                  example: ["Year", "Month"]
                value_col:
                  type: string
                  example: "Sales"
                group_col:
                  type: string
                  example: "Region"
                rolling_window:
                  type: integer
                  default: 10
                  minimum: 2
                  example: 10
                std_multiplier:
                  type: number
                  default: 3
                  minimum: 1
                  maximum: 5
                  example: 3
                resample:
                  type: string
                  enum: [D, W, M, Q, Y]
                  example: "M"
              required:
                - session_id
                - date_col
                - value_col
      responses:
        '200':
          description: Chart generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  chart_data:
                    type: string
                    format: byte
                    description: Base64 encoded image
                  statistics:
                    type: object
                    properties:
                      total_points:
                        type: integer
                      outlier_count:
                        type: integer
                      outlier_percentage:
                        type: number
                      mean:
                        type: number
                      std:
                        type: number
        '400':
          description: Invalid parameters
        '404':
          description: Session not found

  /api/export_ppt/{session_id}:
    get:
      tags:
        - Export
      summary: Export PowerPoint presentation
      description: |
        Generate a professional PowerPoint presentation with all charts from the session.
        Charts are automatically ordered by anomaly count (highest priority first).
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
          description: Session ID from file upload
      responses:
        '200':
          description: PowerPoint file generated
          content:
            application/vnd.openxmlformats-officedocument.presentationml.presentation:
              schema:
                type: string
                format: binary
        '404':
          description: Session not found or no charts available

  /api/export_outliers:
    post:
      tags:
        - Export
      summary: Export outliers to Excel
      description: Export detected outliers to an Excel file with highlighted cells
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                outliers_data:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Excel file with outliers
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary

  /api/health:
    get:
      tags:
        - System
      summary: Health check endpoint
      description: Get application health status and performance metrics
      responses:
        '200':
          description: System is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  uptime:
                    type: string
                  total_requests:
                    type: integer
                  avg_response_time:
                    type: number
                  error_rate:
                    type: number

  /api/statistics:
    get:
      tags:
        - System
      summary: Get usage statistics
      description: Retrieve comprehensive usage statistics and analytics
      responses:
        '200':
          description: Usage statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_uploads:
                    type: integer
                  total_charts:
                    type: integer
                  total_exports:
                    type: integer
                  avg_processing_time:
                    type: number
                  peak_usage_hour:
                    type: integer

components:
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"